Switch to an interface and keep the original version as simple as current behaviour.
