+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/forkchoice/ForkChoiceTrigger.java
@@ -13,79 +13,23 @@
 
 package tech.pegasys.teku.statetransition.forkchoice;
 
-import java.util.concurrent.atomic.AtomicReference;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
-public abstract class ForkChoiceTrigger {
-  private static final Logger LOG = LogManager.getLogger();
-  private final AtomicReference<ForkChoiceUpdate> latestCompletedForkChoice =
-      new AtomicReference<>();
-  private final ForkChoice forkChoice;
+public interface ForkChoiceTrigger {
 
-  protected ForkChoiceTrigger(final ForkChoice forkChoice) {
-    this.forkChoice = forkChoice;
-  }
-
-  public static ForkChoiceTrigger create(
+  static ForkChoiceTrigger create(
       final ForkChoice forkChoice, final boolean balanceAttackMitigationEnabled) {
     return balanceAttackMitigationEnabled
         ? new BalanceAttackMitigationForkChoiceTrigger(forkChoice)
         : new OriginalForkChoiceTrigger(forkChoice);
   }
 
-  public void onSlotStartedWhileSyncing(final UInt64 nodeSlot) {
-    processHead(nodeSlot);
-  }
+  void onSlotStartedWhileSyncing(final UInt64 nodeSlot);
 
-  public abstract void onSlotStarted(final UInt64 nodeSlot);
+  void onSlotStarted(final UInt64 nodeSlot);
 
-  public abstract void onAttestationsDueForSlot(final UInt64 nodeSlot);
+  void onAttestationsDueForSlot(final UInt64 nodeSlot);
 
-  public SafeFuture<Void> requireForkChoiceCompleteForSlot(final UInt64 slot) {
-    final ForkChoiceUpdate forkChoiceUpdate = processHead(slot);
-    if (forkChoiceUpdate.nodeSlot.isGreaterThan(slot)) {
-      return SafeFuture.COMPLETE;
-    } else if (forkChoiceUpdate.nodeSlot.equals(slot)) {
-      return forkChoiceUpdate.result;
-    } else {
-      // Only possible if processHead messed up somehow
-      return SafeFuture.failedFuture(
-          new IllegalStateException(
-              "Requested fork choice be processed for slot "
-                  + slot
-                  + " but result indicates fork choice was only up to "
-                  + forkChoiceUpdate.nodeSlot));
-    }
-  }
-
-  protected ForkChoiceUpdate processHead(final UInt64 nodeSlot) {
-    // Keep trying to get our slot processed until we or someone else gets it done
-    while (true) {
-      final ForkChoiceUpdate previousUpdate = latestCompletedForkChoice.get();
-      if (previousUpdate != null && previousUpdate.nodeSlot.isGreaterThanOrEqualTo(nodeSlot)) {
-        LOG.debug(
-            "Skipping fork choice update for slot {} as high water mark is already {}",
-            nodeSlot,
-            previousUpdate.nodeSlot);
-        return previousUpdate;
-      }
-      final ForkChoiceUpdate newUpdate = new ForkChoiceUpdate(nodeSlot);
-      if (latestCompletedForkChoice.compareAndSet(previousUpdate, newUpdate)) {
-        forkChoice.processHead(nodeSlot).propagateTo(newUpdate.result);
-        return newUpdate;
-      }
-    }
-  }
-
-  private static class ForkChoiceUpdate {
-    private final UInt64 nodeSlot;
-    private final SafeFuture<Void> result = new SafeFuture<>();
-
-    private ForkChoiceUpdate(final UInt64 nodeSlot) {
-      this.nodeSlot = nodeSlot;
-    }
-  }
+  SafeFuture<Void> prepareForBlockProduction(final UInt64 slot);
 }
